Replace SkinImporter and OSZParser with FileUtils.extractZip()
